# specs/openapi.yaml
openapi: 3.0.3
info:
  title: Enterprise AlasGlobalTour API
  version: 1.0.0
  description: >
    API oficial de la plataforma ALAS Global Tour. Gestiona circuitos, eventos,
    competidores, inscripciones, pagos, rankings y contenido editorial.
    Todos los endpoints que expongan puntajes de SurfScores deben incluir
    la atribución "Results by SurfScores.com" en la respuesta.
  contact:
    email: dev@alasglobaltour.com
  license:
    name: Proprietary
servers:
  - url: https://api.alasglobaltour.com/v1
    description: Producción
  - url: https://api-staging.alasglobaltour.com/v1
    description: Staging
components:
  securitySchemes:
    bearerAuth: # Este es un nombre arbitrario
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Introduce el token JWT de Firebase aquí."
security:
  - bearerAuth: []
...
paths:
...
  /inscriptions:
    get:
      tags: [Inscriptions]
      summary: Listar inscripciones (admin)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: eventId
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/InscriptionStatusAdmin'
      responses:
        '200':
          description: Lista de inscripciones
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminInscriptionListResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Inscriptions]
      summary: Crear inscripción (paso 3 del wizard de inscripción)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InscriptionRequest'
      responses:
        '201':
          description: Inscripción creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InscriptionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Cupo agotado o el competidor ya está inscrito en esta categoría
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /inscriptions/{inscriptionId}:
    parameters:
      - $ref: '#/components/parameters/InscriptionId'
    get:
      tags: [Inscriptions]
      summary: Obtener inscripción por ID
      responses:
        '200':
          description: Inscripción encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InscriptionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: [Inscriptions]
      summary: Actualizar inscripción (número de camiseta, estado, notas)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InscriptionUpdateRequest'
      responses:
        '200':
          description: Inscripción actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InscriptionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Inscriptions]
      summary: Cancelar inscripción
      responses:
        '204':
          description: Inscripción cancelada
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: No se puede cancelar una inscripción ya completada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ═══════════════════════════ PAYMENTS ═══════════════════════════
...
  /rankings:
    get:
      tags: [Rankings]
      summary: Ranking por categoría y temporada (caché desde SurfScores)
      description: >
        No usar como live heatboard. Los datos provienen de la caché del BFF .NET.
        La respuesta siempre incluye el campo "attribution" con valor "Results by SurfScores.com".
      parameters:
        - name: categoryId
          in: query
          required: true
          schema:
            type: string
            example: open-hombres
        - name: year
          in: query
          schema:
            type: integer
            example: 2026
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Ranking con atribución obligatoria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RankingResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /rankings/categories:
    get:
      tags: [Rankings]
      summary: Categorías disponibles en el ranking con años disponibles
      responses:
        '200':
          description: Categorías disponibles
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        nombre:
                          type: string
                        availableYears:
                          type: array
                          items:
                            type: integer
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ═══════════════════════════ ARTICLES ═══════════════════════════
...
  /surfscores/sync/{circuitId}:
    parameters:
      - $ref: '#/components/parameters/CircuitId'
    post:
      tags: [SurfScores]
      summary: Sincronizar ranking de un circuito con surfscores.com
      description: >
        El BFF autentica contra surfscores.com/api/v1/users/login con credenciales
        almacenadas en servidor (nunca expuestas al cliente Angular).
        Resultado guardado en caché. Prohibido usar como live heatboard.
      responses:
        '200':
          description: Sincronización exitosa
          content:
            application/json:
              schema:
                type: object
                properties:
                  syncedAt:
                    type: string
                    format: date-time
                  recordsUpdated:
                    type: integer
                  circuitCode:
                    type: string
                    example: ALAS-OPEN-26
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          description: Rate limit de SurfScores alcanzado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Error al conectar con la API de SurfScores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

# ═══════════════════════════════════════════════════════════════════
# COMPONENTS
# ═══════════════════════════════════════════════════════════════════
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── PARAMETERS ───────────────────────────────────────────────
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
    CircuitId:
      name: circuitId
      in: path
      required: true
      schema:
        type: string
        example: open26
...
    InscriptionId:
      name: inscriptionId
      in: path
      required: true
      schema:
        type: string
        example: ins-00123
...
  responses:
    BadRequest:
      description: Error de validación en la solicitud
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # ─── SCHEMAS ──────────────────────────────────────────────────
  schemas:
...
    InscriptionStatusAdmin:
      type: string
      enum: [Pagado, Pendiente]
    InscriptionStatusCompetitor:
      type: string
      enum: [confirmado, pendiente, completado]
    PaymentMethodEnum:
      type: string
      enum: [paypal, beach]
...
    InscriptionRequest:
      type: object
      required: [competitorId, eventId, categoryId, paymentMethod, reglamento]
      properties:
        competitorId:
          type: string
        eventId:
          type: string
        categoryId:
          type: string
        shirtNumber:
          type: string
          example: '#23'
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethodEnum'
        reglamento:
          type: boolean
          description: El competidor acepta el reglamento ALAS
    InscriptionUpdateRequest:
      type: object
      properties:
        shirtNumber:
          type: string
        estadoAdmin:
          $ref: '#/components/schemas/InscriptionStatusAdmin'
        notes:
          type: string
    InscriptionResponse:
      type: object
      properties:
        id:
          type: string
          example: ins-00123
        competitor:
          type: object
          properties:
            id:
              type: string
            fullName:
              type: string
            country:
              type: string
        event:
          type: object
          properties:
            id:
              type: string
            nombre:
              type: string
            lugar:
              type: string
        category:
          type: object
          properties:
            id:
              type: string
            nombre:
              type: string
        circuit:
          type: object
          properties:
            id:
              type: string
            nombre:
              type: string
        shirtNumber:
          type: string
          nullable: true
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethodEnum'
        montoUsd:
          type: number
          format: float
        estadoAdmin:
          $ref: '#/components/schemas/InscriptionStatusAdmin'
        estadoCompetidor:
          $ref: '#/components/schemas/InscriptionStatusCompetitor'
        resultado:
          type: string
          nullable: true
          example: 3° lugar · 920 pts
        transaccionId:
          type: string
          nullable: true
        inscripcionAt:
          type: string
          format: date-time
...
    AdminInscriptionRow:
      type: object
      properties:
        id:
          type: string
        sequentialNumber:
          type: string
          description: Número secuencial de 3 dígitos (e.g., "001")
          example: '007'
        fullName:
          type: string
        country:
          type: string
        ranking2025:
          type: string
          nullable: true
          example: '3°'
        ranking2026:
          type: string
          nullable: true
        categoria:
          type: string
        inscripcionDate:
          type: string
          format: date
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethodEnum'
        montoUsd:
          type: number
          format: float
        estadoAdmin:
          $ref: '#/components/schemas/InscriptionStatusAdmin'
        federacion:
          type: string
        licenciaNumber:
          type: string
        transaccionId:
          type: string
          nullable: true
        notas:
          type: string
          nullable: true
    AdminInscriptionListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AdminInscriptionRow'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'

    # ── PAYMENTS ──────────────────────────────────────────────────
...
    RankingEntry:
      type: object
      properties:
        pos:
          type: integer
          example: 1
        name:
          type: string
        country:
          type: string
          example: Perú
        points:
          type: integer
          example: 9860
        events:
          type: integer
        variation:
          type: integer
          nullable: true
          description: Variación de posición vs. semana anterior (positivo = subió)
    RankingResponse:
      type: object
      properties:
        categoryId:
          type: string
        categoryName:
          type: string
        year:
          type: integer
        cachedAt:
          type: string
          format: date-time
        attribution:
          type: string
          example: Results by SurfScores.com
          description: Campo obligatorio por contrato con SurfScores; debe mostrarse en toda UI que presente estos datos
        data:
          type: array
          items:
            $ref: '#/components/schemas/RankingEntry'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'

    # ── POINTS HISTORY ────────────────────────────────────────────
...
    PaginationMeta:
      type: object
      properties:
        currentPage:
          type: integer
        totalPages:
          type: integer
        totalItems:
          type: integer
        itemsPerPage:
          type: integer
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
        fields:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        timestamp:
          type: string
          format: date-time